ci: test example modules with AddressSanitizer
This requires a certain degree of compatibility between the Rust and C
toolchains, so we use AlmaLinux packages of clang and rustc that are
built against the same LLVM and are known to be compatible.

This also requires unstable compiler options, but `RUSTC_BOOTSTRAP=1`
can be used to force enable nightly features on stable toolchain.