fix: use NgxStr in NgxListIterator to prevent UTF-8 panics
ngx::http::NgxListIterator ngx_str_t items are now NgxStr, not str.

The ngx_str_items in the header name and value are often untrusted
input, and may not have utf-8 contents. The use of ngx_str_t::to_str
in this iterator will panic when the contents are not utf-8. So, instead
of yielding a pair of strs here, yield a pair of &NgxStr, which is like
ngx_str_t but with more methods.